+++ src/index.js
@@ -20,7 +20,6 @@ async function mealCounter() {
   fetch('https://www.themealdb.com/api/json/v1/1/filter.php?c=Seafood')
     .then((response) => response.json())
     .then((data) => {
-      console.log(data)
       linkOne.textContent = `Seafood(${data.meals.length})`;
     });
 }
